requirements.txt
# Implicit dependencies (optional dependencies of dependencies)
django-appconf

# Explicit dependencies (references in code)
django-crispy-forms==1.4.0
django-email-obfuscator
pygeoip==0.3.2
https://github.com/zestedesavoir/Python-ZMarkdown/archive/2.6.0-zds.7.zip

# Api dependencies
djangorestframework==3.1.1
djangorestframework-xml==1.0.1
django-filter==0.9.2
django-oauth-toolkit==0.8.1
drf-extensions==0.2.7
django-rest-swagger==0.2.9
django-cors-headers==1.0.0
